+++ changelog.md
@@ -43,6 +43,7 @@
    4. test loading mask during invoice sorting
    5. test error during invoice sorting
    6. test invoice pagination
+   7. test loading client/invoice page from url parameter
 
 # 2022 03 26 
 - Fix bug when adding client/invoice and total of records not updated in all cached pages